requirements.txt
# python 3.11.9
langchain==0.3.9
langchain_community==0.3.9
langchain_core==0.3.21
langchain_groq==0.2.1
langchain_google_genai==2.0.6
python-dotenv==1.0.1
pydantic==2.10.3
ipykernel
faiss-cpu
pypdf
beautifulsoup4
fastapi
uvicorn
langserve
streamlit
